+++ web-frontend/modules/database/realtime.js
@@ -271,20 +271,21 @@ export const registerRealtimeEvents = (realtime) => {
     const view = store.getters['view/get'](data.view_id)
     if (view !== undefined) {
       if (store.getters['view/getSelectedId'] === view.id) {
-        await store.dispatch('view/forceUpdate', {
+        const updateViewPromise = store.dispatch('view/forceUpdate', {
           view,
           values: data.view,
           repopulate: true,
         })
+        const updateFieldsPromise = store.dispatch('field/forceSetFields', {
+          fields: data.fields,
+        })
+
+        // This makes sure both dispatches are executed in parallel.
+        await Promise.all([updateViewPromise, updateFieldsPromise])
 
         app.$bus.$emit('table-refresh', {
           tableId: store.getters['table/getSelectedId'],
           includeFieldOptions: true,
-          async callback() {
-            await store.dispatch('field/forceSetFields', {
-              fields: data.fields,
-            })
-          },
         })
       }
     }